refactor: move aminoRegistry to messages/ and add export
