+++ packages/proto/src/messages/erc20/aminoRegistry.spec.ts
@@ -1,7 +1,7 @@
 import { createERC20AminoConverters } from './aminoRegistry'
-import { MsgConvertCoin, MsgConvertERC20 } from './tx'
-import { createAminoConverter } from '../../../amino/objectConverter'
-import { expectEqualDefaultAminoConverters } from '../../../../testutils/compareAminoRegistry'
+import { MsgConvertCoin, MsgConvertERC20 } from '../../proto/evmos/erc20/tx'
+import { createAminoConverter } from '../../amino/objectConverter'
+import { expectEqualDefaultAminoConverters } from '../../../testutils/compareAminoRegistry'
 
 describe('test erc20 amino converters', () => {
   it('creates expected converter for MsgConvertCoin', () => {
