+++ packages/proto/src/messages/erc20/aminoRegistry.ts
@@ -1,6 +1,6 @@
 import { AminoConverters } from '@cosmjs/stargate'
-import { MsgConvertCoin, MsgConvertERC20 } from './tx.js'
-import { createAminoConverter } from '../../../amino/objectConverter.js'
+import { MsgConvertCoin, MsgConvertERC20 } from '../../proto/evmos/erc20/tx.js'
+import { createAminoConverter } from '../../amino/objectConverter.js'
 
 export function createERC20AminoConverters(): AminoConverters {
   return {
