+++ packages/proto/src/messages/erc20/index.ts
@@ -2,3 +2,4 @@ export * from './msgConvertCoin.js'
 export * from './msgConvertERC20.js'
 export * from './msgRegisterERC20.js'
 export * from './msgRegisterCoin.js'
+export * from './aminoRegistry.js'
